requirements.txt
langchain>=0.3.25
langchain-openai>=0.3.16
openai>=1.77.0
feedparser>=6.0.10
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.11.18
beautifulsoup4>=4.12.2
tavily-python>=0.7.1 